Reduce some duplicated code in the tests
